+++ CHANGELOG.md
@@ -5,6 +5,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [2.4.0] - 2025-08-25
+
+### Added
+- Added Together AI LLMs and Embedding Models support.
+
+### Changed
+- Changed the sentence-transformers embeddings implementation to handle 'convert_to_tensor' and 'convert_to_numpy' arguments.
+
 ## [2.3.0] - 2025-08-24
 
 ### Added